Rename packages from plural to singular
