+++ src/main/java/rubberduke/task/Deadline.java
@@ -1,4 +1,4 @@
-package rubberduke.tasks;
+package rubberduke.task;
 
 import rubberduke.UserException;
